% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breusch.pagan.test.R
\name{breusch.pagan.test}
\alias{breusch.pagan.test}
\title{Breusch-Pagan Test for Heteroskedasticity}
\usage{
breusch.pagan.test(model, Koenker = TRUE, summary = FALSE)
}
\arguments{
\item{model}{model to be tested fitted "lm" object}

\item{Koenker}{The LM version of Breusch-Pagan don't lie on normality assumption of résiduals}

\item{summary}{Logical. If TRUE, prints summary statistics of the fitted auxiliary regression}
}
\value{
A list with class "htest" containing:
\itemize{
\item statistic: The BP test statistic
\item parameter: Degrees of freedom
\item method: Description of the test method used
\item p.value: P-value of the test
\item null.value: Statement of the null hypothesis
}
}
\description{
Breusch-Pagan Test for Heteroskedasticity
}
\examples{
set.seed(123)
y <- rgamma(107,shape = 321, scale = 6)
X <- matrix(rexp(321, rate = 6)*1000, ncol = 3)
reg1 <- lm(y~X)
breusch.pagan.test(reg1)

}
