% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBER.recession.R
\name{clearConsole}
\alias{clearConsole}
\title{Clear Console}
\usage{
clearConsole()
}
\value{
None
}
\description{
Clears the R console output in both RStudio and other R environments.
}
\examples{
clearConsole()
}
