% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPNGDP.R
\docType{data}
\name{JPNGDP}
\alias{JPNGDP}
\title{Japan Real GDP}
\format{
A quarterly \code{data.frame} with 2 columns:
\describe{
\item{date}{Date of observation (first day of quarter)}
\item{RGDP}{Real GDP in billions of chained 2015 yen}
}
}
\source{
Board of Governors of the Federal Reserve System (US) via FRED
\url{https://fred.stlouisfed.org/series/JPNRGDPEXP}
}
\usage{
JPNGDP
}
\description{
Real Gross Domestic Product for Japan, seasonally adjusted.
}
\examples{
data(JPNGDP)
head(JPNGDP)
plot(JPNGDP$date, JPNGDP$RGDP, type = "l")
}
\keyword{datasets}
