% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss.atest.R
\name{kpss.enhanced.test}
\alias{kpss.enhanced.test}
\title{Kwiatkowski-Phillips-Schmidt-Shin Test for Stationarity}
\usage{
kpss.enhanced.test(
  y,
  type = c("level", "trend"),
  bandwidth.method = c("Newey-West", "KPSS")
)
}
\arguments{
\item{y}{is a vector of time serie}

\item{type}{according to KPSS (1992) the stationary test rely on the regression of y on intercept (level) or on both intercept and trend}

\item{bandwidth.method}{Method to compute bandwidth. Options:
\itemize{
\item "Newey-West": Use Newey-West automatic bandwidth selection from sandwich package
\item "KPSS": Use original KPSS paper's bandwidth selection method
}}
}
\value{
the return object differ from the satndard htest instead, i get inspered by the Eviews format.
}
\description{
Implements an enhanced version of the Kwiatkowski-Phillips-Schmidt-Shin (KPSS) test for stationarity.
This implementation offers flexibility in bandwidth selection and computes the long-run variance
following both the original KPSS paper methodology and using the Newey-West bandwidth selection.
The null hypothesis is that the series is stationary.
This approche differ from the one used by other package which instead use trunc or floor method.
}
