% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagsDiff.R
\name{lagsDiff}
\alias{lagsDiff}
\title{Create differences, logarithmic differences, and lags for a given dataset}
\usage{
lagsDiff(
  dataset,
  diff.vars = NULL,
  lag.vars = NULL,
  diff.log.vars = NULL,
  max.lags = 1,
  max.differences = 1
)
}
\arguments{
\item{dataset}{A dataframe containing the variables to transform}

\item{diff.vars}{A character vector of variable names for which to compute simple differences.}

\item{lag.vars}{A character vector of variable names for which to generate lags.}

\item{diff.log.vars}{A character vector of variable names for which to compute logarithmic differences.}

\item{max.lags}{A positive integer specifying the maximum number of lags to generate, defaut is set to 1}

\item{max.differences}{A positive integer specifying the maximum order of the difference, defaut is set to 1}
}
\value{
A dataframe with the newly added columns
}
\description{
This function adds columns to a dataframe to compute
}
\examples{
#
lagsDiff(longley, diff.vars = "GNP", diff.log.vars = "Unemployed")
}
