% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljung.box.test.R
\name{Ljung.Box.test}
\alias{Ljung.Box.test}
\title{Ljung-Box Test}
\usage{
Ljung.Box.test(x, lags = 10, format = NULL, round = 3)
}
\arguments{
\item{x}{vector of numerical time series}

\item{lags}{maximum lag}

\item{format}{Output format for the table: "rst", "pandoc" or "latex". If NULL, defaults to "rst"}

\item{round}{numeric value for rounding, default is set to 3}
}
\value{
a knitr_kable formatted table with Ljung-Box test results
}
\description{
This function computes the Ljung-Box test statistic for examining the null hypothesis of no autocorrelation in a given time series
}
