% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBER.recession.R
\name{NBER.recession}
\alias{NBER.recession}
\title{Get NBER recession dates}
\usage{
NBER.recession(return_dataframe = TRUE)
}
\arguments{
\item{return_dataframe}{Logical, if TRUE returns data, if FALSE returns URL}
}
\value{
Data frame or URL string
}
\description{
Get NBER recession dates
}
\examples{
NBER.recession()
NBER.recession(FALSE)



}
