% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDiagnostics.R
\name{tsDiagnostics}
\alias{tsDiagnostics}
\title{Plot Ts, ACF and PACF figures}
\usage{
tsDiagnostics(y, x = NULL, lags = 20)
}
\arguments{
\item{y}{Vector of numerical values}

\item{x}{Vector of datetime, default is NULL}

\item{lags}{Maximum number of lags to compute, default is 20}
}
\value{
A combined plot of geom_line_plot, ACF and PACF using patchwork and ADF test results
}
\description{
Plot Ts, ACF and PACF figures
}
\examples{
# Example with date and vector
date <- seq.Date(from = as.Date("1997-12-19"), to =as.Date(Sys.Date()), by = "day")
value <- rnorm(length(date))
tsDiagnostics(x=date, y=value)

# Example with only vector
set.seed(1)
dataNA <- c(NA,rnorm(54321))
tsDiagnostics(y=dataNA)

}
