% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white.test.R
\name{white.test}
\alias{white.test}
\title{White Test for Heteroskedasticity}
\usage{
white.test(model, summary = FALSE)
}
\arguments{
\item{model}{fitted model object (lm, dynlm, ardl, etc.)}

\item{summary}{logical. If set to TRUE summary statistics of the fitted linear model is printed}
}
\value{
An object of class "htest" containing the test statistic and p-value.
}
\description{
White Test for Heteroskedasticity
}
