% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljung.box.test.R
\name{ljung.box.test}
\alias{ljung.box.test}
\title{Ljung-Box Test}
\usage{
ljung.box.test(x, lag = 10, format = NULL, round = 3)
}
\arguments{
\item{x}{vector of numerical time series}

\item{lag}{maximum lag}

\item{format}{Output format for the table: "rst", "pandoc" or "latex". If NULL, defaults to "rst"}

\item{round}{numeric vector for round, default is set to 3}
}
\value{
a knitr_kable formatted table with descriptive statistics
}
\description{
This function compute the Ljung-Box test statistic for examining the null hypothesis of no autocorrelation in a given time series
}
\examples{
set.seed(1)
tsatoolkit::ljung.box.test(rnorm(100))

}
